requirements.txt
asgiref==3.6.0
# backports.zoneinfo==0.2.1
backports.zoneinfo==0.2.1; python_version < "3.9"
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cmake==3.26.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
defusedxml==0.7.1
Django==4.1.6
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==22.1
django-phonenumber-field==7.0.2
django-sslserver==0.22
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
dlib==19.24.1
f==0.0.1
fabric==3.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==3.4
importlib-metadata==7.1.0
invoke==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
Markdown==3.6
MarkupSafe==2.1.2
numpy==1.24.3
oauthlib==3.2.2
pandas==2.0.1
paramiko==3.0.0
phonenumbers==8.13.5
Pillow==9.4.0
prettytable==3.6.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pyOpenSSL==23.1.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
threadpoolctl==3.1.0
typing-extensions==4.11.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.3.4
zipp==3.18.1
